requirements.txt
Flask
tensorflow
Pillow
numpy
Werkzeug
gunicorn